Delete nci_script.sh
This file is being replaced by a new 'process_pyrate_xxx' template file. The new template name is consistent with how GAMMA is run on the NCI. 

In addition, the modules listed are to move to a central PYRATE_CONFIG file that's been created on the NCI. The environment variables are to be added to a user's .profile and .bashrc files when they setup running PyRate for the first time.