Update mst to work with networkx v2.
Issue was that .edges() (and nodes etc) have changed from being list/dict copies to nx's own View objects. There was a guard checking if the mst edges was a list instance before iterating over them. Changing this guard to check if it's an instance of EdgeView fixes the problem.